08-web-scraper-project/requirements.txt
pymongo==4.7.1
dnspython==2.6.1
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
python-dotenv==1.0.1
